refactor for aggregation clases and mysql sink fixes
